Moved ChunkData to its own module
