+++ amulet/level/formats/leveldb_world/dimension.py
@@ -8,19 +8,18 @@ from typing import (
     List,
     TYPE_CHECKING,
     Tuple,
-    Union,
-    Iterable,
 )
 from threading import RLock
 import logging
 from contextlib import suppress
 
 import amulet_nbt
-from amulet_nbt import TAG_Long, TAG_String, TAG_Compound, NBTFile, NBTLoadError
+from amulet_nbt import TAG_Long, TAG_String, TAG_Compound, NBTLoadError
 
 from amulet.api.errors import ChunkDoesNotExist, DimensionDoesNotExist
 from amulet.api.data_types import ChunkCoordinates
 from amulet.libs.leveldb import LevelDB
+from .chunk import ChunkData
 
 if TYPE_CHECKING:
     from amulet.api.data_types import Dimension
@@ -74,37 +73,6 @@ class ActorCounter:
             count = self._count
             self._count += 1
         return self._session, count
-
-
-class ChunkData(Dict[bytes, bytes]):
-    def __init__(
-        self,
-        chunk_data: Union[Dict[bytes, bytes], Iterable[Tuple[bytes, bytes]]] = (),
-        *,
-        entity_actor: Iterable[NBTFile] = (),
-        unknown_actor: Iterable[NBTFile] = (),
-    ):
-        super().__init__(chunk_data)
-        self._entity_actor = list(entity_actor)
-        self._unknown_actor = list(unknown_actor)
-
-    @property
-    def entity_actor(self) -> List[NBTFile]:
-        """
-        A list of entity actor data.
-        UniqueID is stripped out. internalComponents is stripped out if there is no other data.
-        """
-        return self._entity_actor
-
-    @property
-    def unknown_actor(self) -> List[NBTFile]:
-        """
-        A list of actor data that does not fit in the known actor types.
-        UniqueID is stripped out.
-        internalComponents.{StorageKeyComponent}.StorageKey is replaced with a blank string.
-        All keys matching this pattern will get replaced with the real storage key when saving (at least one must exist)
-        """
-        return self._unknown_actor
 
 
 class LevelDBDimensionManager:
